requirements.txt
fastapi
uvicorn
langchain
langchain-community
sentence-transformers
faiss-cpu
python-dotenv
pydantic
pytest
requests
langgraph
google-generativeai
